Sending Special Keys - Selenium
If the cursor is not currently in a text field, pressing the HOME and END keys will scroll the browser to the top and bottom of the page, respectively. The send_keys() calls scroll the page. Calling browser.find_element_by_tag_name('html') is a good place to send keys to the general web page.